improved the structure for selenium test of APP
