+++ oscm-portal-integrationtests/pom.xml
@@ -33,6 +33,7 @@
           <execution>
             <phase>test</phase>
         <configuration>
+        <skipTests>true</skipTests>
           <redirectTestOutputToFile>true</redirectTestOutputToFile>
           <reportsDirectory>../shippable/testresults</reportsDirectory>
         </configuration>
