+++ oscm-webtest-base/src/main/java/org/oscm/webtest/AppHtmlElements.java
@@ -1,0 +1,12 @@
+package org.oscm.webtest;
+
+public class AppHtmlElements{
+    
+    public static final String APP_TITLE = "APP Controller Configuration";
+    
+    public static final String APP_FORM_CONFIGURATIONSETTING = "configurationSettings";
+    public static final String APP_INPUT_END_NEWCONTROLLERID = ":newControllerId";
+    public static final String APP_INPUT_END_NEWORGID = ":newOrgnizationId";
+    public static final String APP_INPUT_END_EXISTORGID = ":configurationValue";
+    
+}
